add a deadman switch for kazoo reconnections
In tests, if we firewall off the zookeeper port, kazoo
just infinitely tries to reconnect and never “loses” the
connection.

In theory, split brain should be avoided by the appservers
using the new master, but we should really make sure that this
master is shut down timeously.

This commit ensures that