make restarting more robust
remove “_stopping” attribute and checks, this was
a bad idea anyway as restart is synchronous and
always called the main thread. So multiple calls
are not a problem.

Also call ._stop() higher the method so
exceptions in the following code do not prevent
an asyncio exit

Do more logging, so if there is a problem in future
it can be pinpointed

A pending challenge is what to do if something in
the restart method hangs for a long time or forever.
In that case